# Add support for other char types, opt-in SIMD, JSON parse rules