--- a/include/fast_float/fast_float.h
+++ b/include/fast_float/fast_float.h
@@ -27,7 +27,7 @@ struct parse_options {
   constexpr explicit parse_options(
       chars_format fmt = chars_format::general,
       parse_rules rules = parse_rules::std_rules,
-      bool parse_ints = false, char dot = '.', )
+      bool parse_ints = false, char dot = '.')
     : format(fmt), rules(rules), parse_ints(parse_ints), decimal_point(dot) {}
 
   /** Which number formats are accepted */