- Add support for other char types (with opt-in SSE2 support to accelerate parsing)
- Add JSON parse rules support