@lemire 2023-05-01
@mayawarrier  Can you sync with the main branch?

Also I recommend breaking down the PR. If you want to extend for author syntaxes like JSON, please do so as a separate PR from SIMD optimizations... Otherwise it gets too difficult to review.

Make the PR as small as you can.