@mayawarrier 2023-05-02
@lemire I've synced with the main branch and reduced the PR to just the SIMD optimizations, as requested. Awaiting your review!