@lemire 2023-05-02
I am concerned that currently, in this PR, the SSE2 code does not get compiled at all...