@mayawarrier 2023-05-02
> I am concerned that currently, in this PR, the SSE2 code does not get compiled at all...

I've added automatic detection of SSE2, so hopefully it should be compiled and tested now.