@lemire 2023-05-02
@mayawarrier

Good. Let us run tests. If/when it passes, I will need to do some hand inspection because it has to do with performance and I must make sure we are not messing anything.