@lemire 2023-05-05
On my todo to review. I will eventually merge this.