@mayawarrier 2023-05-07
@lemire  I made a small perf change. Can you run the tests again?