@lemire 2023-05-08
Running tests.

Note that I still lack benchmarking for this, which I will need to add.

I won't add an optimization without some quantification of the benefits.