@mayawarrier 2023-05-09
@lemire I attempted to benchmark this using a modified version of the simple_fastfloat_benchmark repo, this was one of the runs on Linux (WSL) using gcc 11.3:

![benchmark_results_cropped](https://github.com/fastfloat/fast_float/assets/34803055/e9b163a1-5ed3-4d52-b40c-abde09f1476f)
I'm seeing on average 10-15% increase in speed for the char16_t SIMD version.

With MSVC however the SIMD version shows a decrease in speed:

![benchmark_results2_cropped](https://github.com/fastfloat/fast_float/assets/34803055/7bb0351a-8703-4836-8e5e-2afe3ecb2493)