@lemire 2023-05-10
Ok. I have effectively merged your code changes into https://github.com/lemire/simple_fastfloat_benchmark (I did not quite do a PR because I wanted to preserve the project)....

I went to an Ice Lake server (GCC11). I did this...

```
cmake -B build
cmake --build build
./build/benchmarks/benchmark -f data/canada.txt 
```

And I got...
```
# read 111126 lines 
ASCII volume = 1.93374 MB 
netlib                                  :   251.28 MB/s (+/- 29.4 %)    14.44 Mfloat/s      30.55 i/B   557.46 i/f (+/- 0.0 %)     12.01 c/B   219.14 c/f (+/- 1.8 %)      2.54 i/c      3.16 GHz 
doubleconversion                        :   224.78 MB/s (+/- 0.5 %)    12.92 Mfloat/s      54.10 i/B   987.16 i/f (+/- 0.0 %)     13.52 c/B   246.61 c/f (+/- 0.6 %)      4.00 i/c      3.19 GHz 
strtod                                  :   160.91 MB/s (+/- 0.7 %)     9.25 Mfloat/s      70.85 i/B  1292.81 i/f (+/- 0.0 %)     18.88 c/B   344.55 c/f (+/- 0.7 %)      3.75 i/c      3.19 GHz 
abseil                                  :   409.27 MB/s (+/- 2.3 %)    23.52 Mfloat/s      29.36 i/B   535.72 i/f (+/- 0.0 %)      7.43 c/B   135.53 c/f (+/- 2.2 %)      3.95 i/c      3.19 GHz 
fastfloat                               :   871.76 MB/s (+/- 2.3 %)    50.10 Mfloat/s      16.55 i/B   302.00 i/f (+/- 0.0 %)      3.49 c/B    63.62 c/f (+/- 2.3 %)      4.75 i/c      3.19 GHz 
UTF-16 volume = 3.86749 MB 
doubleconversion                        :   461.12 MB/s (+/- 1.9 %)    13.25 Mfloat/s      27.11 i/B   989.16 i/f (+/- 0.0 %)      6.59 c/B   240.47 c/f (+/- 1.9 %)      4.11 i/c      3.19 GHz 
fastfloat                               :  1382.45 MB/s (+/- 1.2 %)    39.72 Mfloat/s      10.34 i/B   377.51 i/f (+/- 0.0 %)      2.20 c/B    80.22 c/f (+/- 1.2 %)      4.71 i/c      3.19 GHz 
```

I did changed the `CMakeLists.txt` file as follows...

```diff
diff --git a/CMakeLists.txt b/CMakeLists.txt
index f72f6a6..980fc4b 100644
--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -30,7 +30,7 @@ endif()
 
 
 FetchContent_Declare(fast_float
-     GIT_REPOSITORY https://github.com/fastfloat/fast_float.git
+     GIT_REPOSITORY https://github.com/mayawarrier/fast_float
      GIT_TAG origin/main
      GIT_SHALLOW TRUE)
```

And I did...

```
cmake -B buildnew
cmake --build buildnew
./buildnew/benchmarks/benchmark -f data/canada.txt 
```

and I got...

```
# read 111126 lines 
ASCII volume = 1.93374 MB 
netlib                                  :   251.94 MB/s (+/- 27.6 %)    14.48 Mfloat/s      30.55 i/B   557.46 i/f (+/- 0.0 %)     12.05 c/B   219.95 c/f (+/- 1.4 %)      2.53 i/c      3.18 GHz 
doubleconversion                        :   223.22 MB/s (+/- 0.6 %)    12.83 Mfloat/s      54.10 i/B   987.16 i/f (+/- 0.0 %)     13.62 c/B   248.58 c/f (+/- 0.5 %)      3.97 i/c      3.19 GHz 
strtod                                  :   161.15 MB/s (+/- 11.3 %)     9.26 Mfloat/s      70.85 i/B  1292.81 i/f (+/- 0.0 %)     18.87 c/B   344.37 c/f (+/- 11.2 %)      3.75 i/c      3.19 GHz 
abseil                                  :   404.10 MB/s (+/- 1.2 %)    23.22 Mfloat/s      29.36 i/B   535.72 i/f (+/- 0.0 %)      7.52 c/B   137.16 c/f (+/- 1.2 %)      3.91 i/c      3.19 GHz 
fastfloat                               :   870.16 MB/s (+/- 1.3 %)    50.01 Mfloat/s      16.51 i/B   301.21 i/f (+/- 0.0 %)      3.49 c/B    63.67 c/f (+/- 1.3 %)      4.73 i/c      3.18 GHz 
UTF-16 volume = 3.86749 MB 
doubleconversion                        :   430.57 MB/s (+/- 0.6 %)    12.37 Mfloat/s      27.11 i/B   989.16 i/f (+/- 0.0 %)      7.06 c/B   257.71 c/f (+/- 0.5 %)      3.84 i/c      3.19 GHz 
fastfloat                               :  1333.69 MB/s (+/- 0.6 %)    38.32 Mfloat/s      10.38 i/B   378.62 i/f (+/- 0.0 %)      2.28 c/B    83.16 c/f (+/- 0.6 %)      4.55 i/c      3.19 GHz 
```

It is not very conclusive.

The `canada.txt` file should be a good test for this because it has many long sequences.

Admittedly, maybe I screwed up somewhere...