@lemire 2023-05-12
In one instance, you are reporting performance counters (instructions per byte and so forth) and in the other instance you are not. WSL won't support performance counters (at least I never managed to get it to do it with me), which is problematic here because the reason SIMD ought to improve the performance has to do with reduced instruction count.

I do not see a reduction in the number of instructions. So something is off.

Here is the same thing on an older processor (AMD Zen2) with GCC11 (Ubuntu 22).

Main:

```
$ ./build/benchmarks/benchmark -f data/canada.txt
# read 111126 lines 
ASCII volume = 1.93374 MB 
netlib                                  :   250.62 MB/s (+/- 0.7 %)    14.40 Mfloat/s      31.21 i/B   569.44 i/f (+/- 0.0 %)     12.89 c/B   235.20 c/f (+/- 0.6 %)      2.42 i/c      3.39 GHz 
doubleconversion                        :   182.96 MB/s (+/- 1.7 %)    10.51 Mfloat/s      56.01 i/B  1021.94 i/f (+/- 0.0 %)     17.64 c/B   321.94 c/f (+/- 1.7 %)      3.17 i/c      3.38 GHz 
strtod                                  :   136.55 MB/s (+/- 1.6 %)     7.85 Mfloat/s      69.74 i/B  1272.61 i/f (+/- 0.0 %)     23.65 c/B   431.45 c/f (+/- 1.6 %)      2.95 i/c      3.39 GHz 
abseil                                  :   375.14 MB/s (+/- 0.3 %)    21.56 Mfloat/s      30.35 i/B   553.72 i/f (+/- 0.0 %)      8.61 c/B   157.13 c/f (+/- 0.2 %)      3.52 i/c      3.39 GHz 
fastfloat                               :   785.45 MB/s (+/- 0.2 %)    45.14 Mfloat/s      16.93 i/B   309.00 i/f (+/- 0.0 %)      4.11 c/B    75.06 c/f (+/- 0.1 %)      4.12 i/c      3.39 GHz 
UTF-16 volume = 3.86749 MB 
doubleconversion                        :   362.47 MB/s (+/- 0.8 %)    10.41 Mfloat/s      28.03 i/B  1022.95 i/f (+/- 0.0 %)      8.91 c/B   325.13 c/f (+/- 0.7 %)      3.15 i/c      3.39 GHz 
fastfloat                               :  1242.90 MB/s (+/- 0.2 %)    35.71 Mfloat/s      10.54 i/B   384.51 i/f (+/- 0.0 %)      2.60 c/B    94.83 c/f (+/- 0.1 %)      4.05 i/c      3.39 GHz 
```

This PR:

```
$ ./buildnew/benchmarks/benchmark -f data/canada.txt
# read 111126 lines 
ASCII volume = 1.93374 MB 
netlib                                  :   259.52 MB/s (+/- 0.5 %)    14.91 Mfloat/s      31.21 i/B   569.44 i/f (+/- 0.0 %)     12.44 c/B   227.06 c/f (+/- 0.4 %)      2.51 i/c      3.39 GHz 
doubleconversion                        :   182.41 MB/s (+/- 0.8 %)    10.48 Mfloat/s      56.01 i/B  1021.94 i/f (+/- 0.0 %)     17.70 c/B   322.95 c/f (+/- 0.8 %)      3.16 i/c      3.39 GHz 
strtod                                  :   138.46 MB/s (+/- 1.0 %)     7.96 Mfloat/s      69.74 i/B  1272.61 i/f (+/- 0.0 %)     23.32 c/B   425.52 c/f (+/- 1.0 %)      2.99 i/c      3.39 GHz 
abseil                                  :   347.22 MB/s (+/- 1.5 %)    19.95 Mfloat/s      30.35 i/B   553.72 i/f (+/- 0.0 %)      9.30 c/B   169.75 c/f (+/- 1.3 %)      3.26 i/c      3.39 GHz 
fastfloat                               :   792.75 MB/s (+/- 0.2 %)    45.56 Mfloat/s      16.89 i/B   308.21 i/f (+/- 0.0 %)      4.07 c/B    74.35 c/f (+/- 0.1 %)      4.15 i/c      3.39 GHz 
UTF-16 volume = 3.86749 MB 
doubleconversion                        :   382.90 MB/s (+/- 0.3 %)    11.00 Mfloat/s      28.03 i/B  1022.95 i/f (+/- 0.0 %)      8.44 c/B   307.84 c/f (+/- 0.2 %)      3.32 i/c      3.39 GHz 
fastfloat                               :  1201.45 MB/s (+/- 1.0 %)    34.52 Mfloat/s      10.57 i/B   385.62 i/f (+/- 0.0 %)      2.69 c/B    98.02 c/f (+/- 1.0 %)      3.93 i/c      3.38 GHz 
```

Please focus on the `i/B` column. If this column does not show gains (i.e., reductions), then the PR is not likely beneficial.